Revert it back to seperate statuslist file
