+++ src/frontend/src/pages/PIAForm/BannerStatus/index.tsx
@@ -1,7 +1,7 @@
 import { IBannerStatusProps } from './interfaces';
 import Callout from '../../../components/common/Callout';
 import { getUserRole } from '../../../utils/user';
-import { statusList } from '../../../utils/statusList/common';
+import { statusList } from '../../../utils/statusList/statusList';
 const BannerStatus = ({ pia }: IBannerStatusProps) => {
   const populateBanner = () => {
     const currentStatus = pia.status || 'Completed';
